Updated Add Crops page to fix layout, styling, errors, etc
