+++ src/dataclients/ProcessesClient.js
@@ -1,0 +1,7 @@
+import { API_URL, handleErrors } from 'utils/FetchUtils';
+
+export const getProcessCategories = async () => {
+  return fetch(API_URL + 'fieldCropCycles/processCategories')
+    .then(handleErrors)
+    .then((response) => response.json());
+};
